+++ app/src/main/java/ComAz007/calculator/Libs/expression/Expression.kt
@@ -1,5 +1,5 @@
 package expression
 
 interface Expression {
-    fun evaluate(): Int
+    fun evaluate(): Float
 }
